requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
Flask-RESTful
PyJWT
Werkzeug
pymysql
mysql-connector-python
python-dotenv
requests
Jinja2
pandas
openpyxl  

